tests: updated test to work on Flask 2.1.x
Big part of tests is initializing Flask app with different instances of JSON:API 
middleware (`flask_rest_jsonapi_next.Api`). For that reason, it was never a good idea
to have `app` fixture scoped to test session: tests are repeatedly "re initializing" this
instance of app.

In newer versions of Flask, this kind of behavior is explicitly forbidden and raises error.

For that reason tests now avoid multiple calls of `init_app` on same instance of Flask app.